Barky/requirements.txt
-e git+https://github.com/ahuimanu/cidm6330.git@9153d230de244ecb5d54a3f4088b6b0afa0ce1cb#egg=allocation&subdirectory=app\src
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
-e git+https://github.com/ahuimanu/cidm6330.git@9153d230de244ecb5d54a3f4088b6b0afa0ce1cb#egg=barkylib&subdirectory=barky\src
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pep517==0.10.0
pip-tools==6.0.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
python-dotenv==0.17.0
redis==3.5.3
regex==2021.4.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.7
tenacity==7.0.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wincertstore==0.2
